requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt[blacklist]>=5.3
python-dotenv>=1.0
psycopg2-binary>=2.9 